==> project1/00-import_and_pkl_raw_data.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from string import Template

def get_MTA_data(startdate, enddate):
    data = pd.DataFrame()
    startdate = pd.to_datetime(startdate)  ### Assumption: Dates are Saturdays
    enddate = pd.to_datetime(enddate)
    
   # url = 'http://web.mta.info/developers/data/nyct/turnstile/turnstile_180623.txt'
    urlbase = Template('http://web.mta.info/developers/data/nyct/turnstile/turnstile_$date.txt')
    
    week = pd.Timedelta('7 days')
    next_saturday = startdate
    
    week_counter = 1
    
    while(next_saturday <= enddate):
        urlday = next_saturday.strftime('%y%m%d')
        url = urlbase.substitute(date=urlday)
        next_saturday = next_saturday+week
        
        newdata = pd.read_csv(url)
        
        #label each week's observations with a different integer
        newdata['week'] = week_counter
        week_counter += 1
        
        data = pd.concat([data,newdata])
        
   # url = urlbase.substitute(date=date)
    return(data)
    




data = get_MTA_data('04/28/2018','06/30/2018')


with open('MTA_Apr_June_with_weeks.pickle', 'wb') as f:
    pickle.dump(data,f)
