==> project1/01-clean_and_get_weekly_totals.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from string import Template
import matplotlib.pyplot as plt
%matplotlib inline

with open('MTA_Apr_June_with_weeks.pickle', 'rb') as f:
    data = pickle.load(f)

#remove white space from columns


data.columns = [column.strip() for column in data.columns]
data.rename(columns={'week':'WEEK'}, inplace=True)
data.head()

#data.info()
data[data['STATION'] == '23 ST'].head()

data['TIME'] = pd.to_datetime(data['TIME'], infer_datetime_format=True, cache=True).apply(lambda x: x.time())
data.TIME.value_counts()

#that's a lot of data! let's drop some observations off the bat. Getting rid of observations from 9:00- 12pm, 
#leaving only 12:00 pm-9:00 pm, when most tech-employees and potential donors might be commuting home or getting lunch

data = data[(data['TIME']>pd.to_datetime('12:00:00').time()) & (data['TIME']<pd.to_datetime('21:00:00').time())]

data.shape

#function to calculate the difference between an observation's max and min values, which will mostly occur at 12:00 and 8:.
#will be applied on unique turnstile_id + day-level
def difference(srs):
    return srs.max() - srs.min()

#Identifying and grouping by unique turnstiles, taking the daily max-daily min for that turnstile:

turnstile_col_names =['STATION','C/A','UNIT','SCP']

by_turnstile_date =  data.groupby(by=turnstile_col_names+['WEEK']+['DATE'], as_index=False).agg(difference)
by_turnstile_date['turnstile_id'] = by_turnstile_date.apply(lambda x: str(x['STATION']+'_'+x['C/A']+'_'+x['UNIT']+'_'+x['SCP']), axis=1)

by_turnstile_date['ENTRIES'].describe()
#plt.hist(by_station_date['ENTRIES'][by_station_date['ENTRIES']>861])

#turnstiles with 1 really high day's entries probably are showing data anomalies:

by_turnstile_date[by_turnstile_date['ENTRIES']>4000].turnstile_id.value_counts()

#
#identify outlier values of 'ENTRIES' for Turnstile/Date

high_turnstile_count = dict(by_turnstile_date[by_turnstile_date['ENTRIES']>4000].turnstile_id.value_counts())
high_turnstile_count = {key : val for key,val in high_turnstile_count.items() if val<2}
high_turnstile_count

#identify outlier values of 'ENTRIES' for turnstile/Date pairs and replace them with weekly averages for the turnstile
avg_turnstile_per_day = by_turnstile_date.groupby(by=['STATION','turnstile_id','WEEK'], as_index=False).mean().drop(columns='EXITS')
avg_turnstile_per_day.rename(columns={'ENTRIES':'AVG_REPLACEMENT'}, inplace=True)

by_turnstile_date_anomalies = by_turnstile_date[(by_turnstile_date['ENTRIES']>5000) & (by_turnstile_date['turnstile_id'].isin(high_turnstile_count.keys()))]
by_turnstile_date_anomalies = by_turnstile_date_anomalies.merge(right=avg_turnstile_per_day, how='left')
by_turnstile_date_anomalies['ENTRIES'] = by_turnstile_date_anomalies['AVG_REPLACEMENT']
by_turnstile_date_anomalies.drop(columns=['AVG_REPLACEMENT'], inplace=True)

by_turnstile_date = pd.concat([by_turnstile_date[(by_turnstile_date['ENTRIES']<=30000) | ~(by_turnstile_date['turnstile_id'].isin(high_turnstile_count.keys()))], by_turnstile_date_anomalies])

by_turnstile_date.ENTRIES.describe()

by_station_date = by_turnstile_date.groupby(by=['STATION','WEEK', 'DATE'], as_index=False).sum()
by_station_date.ENTRIES.describe()

#stations with 1 really high day's entries probably are showing data anomalies.
by_station_date[by_station_date['ENTRIES']>30000].STATION.value_counts()

#identify outlier values of 'ENTRIES' for Station/Date pairs and replace them with weekly averages for the station
high_day_count = dict(by_station_date[by_station_date['ENTRIES']>30000].STATION.value_counts())
high_day_count = {key : val for key,val in high_day_count.items() if val<2}
high_day_count

#identify outlier values of 'ENTRIES' for Station/Date pairs and replace them with weekly averages for the station
avg_day_per_week = by_station_date.groupby(by=['STATION','WEEK'], as_index=False).mean().drop(columns='EXITS')
avg_day_per_week.rename(columns={'ENTRIES':'AVG_REPLACEMENT'}, inplace=True)

by_station_date_anomalies = by_station_date[(by_station_date['ENTRIES']>30000) & (by_station_date['STATION'].isin(high_day_count.keys()))]
by_station_date_anomalies = by_station_date_anomalies.merge(right=avg_day_per_week, how='left')
by_station_date_anomalies['ENTRIES'] = by_station_date_anomalies['AVG_REPLACEMENT']
by_station_date_anomalies.drop(columns=['AVG_REPLACEMENT'], inplace=True)

by_station_date = pd.concat([by_station_date[(by_station_date['ENTRIES']<=30000) | ~(by_station_date['STATION'].isin(high_day_count.keys()))], by_station_date_anomalies])

by_station_date.ENTRIES.describe()

by_station_date.head()

by_station_date.groupby(by=['STATION','WEEK'], as_index=False).count().ENTRIES.unique()

#drop station/week combos with fewer than 7 days per week observed

station_date_counts = by_station_date.groupby(by=['STATION','WEEK'], as_index=False).count()
station_date_counts = station_date_counts[station_date_counts.ENTRIES < 7]
station_date_counts.drop(columns=['ENTRIES','EXITS','DATE'], inplace=True)
by_station_date = by_station_date.merge(right=station_date_counts, on=['STATION','WEEK'], how='left', indicator = True)
by_station_date = by_station_date[by_station_date._merge == 'left_only']
by_station_date.drop(columns=['_merge'], inplace=True)

by_station_date.ENTRIES.describe()

with open('by_station_date_cleaned.pickle','wb') as f:
    pickle.dump(by_station_date,f)

by_station_week =  by_station_date.groupby(by=['STATION', 'WEEK'], as_index=False).sum().drop(columns=['EXITS'])
by_station_week.head()

by_station_week.ENTRIES.describe()

by_station_week[by_station_week['ENTRIES']>500000]

#get rid of anomalously high weeks
by_station_week = by_station_week[by_station_week['ENTRIES']<500000]

weekly_avg_by_station = by_station_week.groupby(by="STATION", as_index=False).mean().drop(columns=['WEEK'])
weekly_avg_by_station["mean_percentage"] = (weekly_avg_by_station["ENTRIES"]/weekly_avg_by_station["ENTRIES"].sum())*100
weekly_avg_by_station = weekly_avg_by_station.sort_values(by='ENTRIES', ascending=False)
weekly_avg_by_station.head(10)

with open('weekly_avg_by_station_clean.pickle', 'wb') as to_write:
    pickle.dump(weekly_avg_by_station, to_write)

weekly_avg_by_station.head(10)

weekly_avg_by_station.head(10).mean_percentage.sum()
