==> project1/03-distance_calculations_v2.ipynb <==
import pandas as pd
import pickle
from haversine import haversine

with open('foursquare_POIs_only.pickle', 'rb') as f:
    fsq_df = pickle.load(f)

fsq_df = fsq_df[['venueId','venueCategory', 'latitude', 'longitude']].rename(columns={'latitude':'poi_lat', 'longitude':'poi_lon'})

fsq_df.shape

#reducing observations to unique venues/POIs for distance calculation
fsq_df_unique = fsq_df.groupby(by='venueId', as_index=False).max()
fsq_df_unique.shape

with open('weekly_avg_by_station_clean.pickle', 'rb') as f:
    entry_df = pickle.load(f)

entry_df = entry_df.sort_values('ENTRIES', ascending = False)

#only interested in top 10 stations
entry_df = entry_df.iloc[0:10]

#only interested in the station names and orders
distance_df = pd.DataFrame(entry_df.STATION)
distance_df.rename(columns={'STATION': 'station'}, inplace=True)

#manually update latitude/longitude for top 10
stat_lat_lons =[(40.750326,-73.988067), # 34 st - Herald Square
                (40.752962,-73.977219), # grand central - 42 st
                (40.750497,-73.990877), #34 st- penn station
                (40.741623,-73.989365), #23 st - actually 5th and 23rd st
                (40.755453,-73.987285), # Times Square - 42nd st
                (40.709641,-74.008283), #Fulton St
                (40.758923,-73.981350), #47-50 STS ROCK 
                (40.735019,-73.990698), # 14 st - Union Square
                (40.768868,-73.980959), #59 St Columbus
                (40.712925,-74.009928)] # PATH NEW WTC
                

distance_df['station_lat'] = [x[0] for x in stat_lat_lons]
distance_df['station_lon'] = [x[1] for x in stat_lat_lons]
distance_df

#add column of all 1s to each df so we can merge to get all distances
distance_df['merge_column'] = 1
fsq_df_unique['merge_column'] = 1

distance_df = distance_df.merge(right=fsq_df_unique, on='merge_column')
distance_df.drop('merge_column', axis=1, inplace=True)

distance_df.head()

def distance_to_station(df_row):
    stat_coords = (df_row.station_lat, df_row.station_lon)
    poi_coords = (df_row.poi_lat, df_row.poi_lon)
    return haversine(stat_coords, poi_coords, miles=True)

distance_df['distance'] = distance_df.apply(distance_to_station, axis=1)

distance_df.head()

distance_df = distance_df[distance_df.distance < 1]
distance_df.info()

with open('distance_df_v2.pickle','wb') as f:
    pickle.dump(distance_df, f)

distance_df.groupby(by=['venueCategory','station'], as_index=False).count().sort_values(['venueCategory','poi_lat'], ascending=[True,False])
