==> project1/04-combined_weighting_calculations.ipynb <==
import pandas as pd
import numpy as np
import pickle
from scipy import stats

with open('distance_df_v2.pickle', 'rb') as f:
    distance_df = pickle.load(f)
distance_df.head()

with open('Benson_full_analysis.pickle_v2', 'rb') as f:
    station_entries_df = pickle.load(f)
station_entries_df.head()

distance_by_station = distance_df.groupby(by='station', as_index=False).count()

distance_df.head()

station_priorities = station_entries_df[['STATION', 'ENTRIES']].rename(columns={'ENTRIES': 'avg_entries', 'STATION': 'station'})
station_priorities.head()

station_priorities = station_priorities.merge(distance_by_station, on='station')

station_priorities.head()

station_priorities['avg_entries'] = station_priorities['avg_entries'].astype(int)

station_priorities.drop(['station_lat', 'station_lon', 'venueId', 'venueCategory', 'poi_lat', 'poi_lon'], axis=1, inplace = True)

station_priorities.rename(columns={'distance':'nearby_poi_count'},inplace=True)

station_priorities['entry_weight'] = station_priorities['avg_entries']/station_priorities['avg_entries'].sum()

station_priorities['poi_weight'] = station_priorities['nearby_poi_count']/station_priorities['nearby_poi_count'].sum()

#station_priorities['combined_weight_product'] = station_priorities['entry_weight'] * station_priorities['poi_weight']
station_priorities['combined_share_mean'] = station_priorities[['entry_weight','poi_weight']].mean(axis=1)
station_priorities.rename(columns={'entry_weight':'entries_share', 'poi_weight':'poi_share'}, inplace=True)
#station_priorities['combined_weight_harmonic_mean'] = stats.hmean(station_priorities[['entry_weight','poi_weight']], axis=1)

station_priorities.sort_values('combined_share_mean', inplace=True, ascending=False)

station_priorities

#distance_df.groupby(by=['venueCategory','station'], as_index=False).count().sort_values(['venueCategory','poi_lat'], ascending=[True,False])

with open('final_station_priorities_v2.pickle', 'wb') as f:
    pickle.dump(station_priorities, f)
